+++ src/main/java/com/app/modules/weather/service/HistoricoPrecipitacionesService.java
@@ -240,7 +240,7 @@ public class HistoricoPrecipitacionesService {
                                         EstacionesDTO estacionesDTO = new EstacionesDTO();
                                         estacionesDTO.setLatitud(nodo.get("latitud").asString());
                                         estacionesDTO.setProvincia(nodo.get("provincia").asString());
-                                        estacionesDTO.setAltitud(nodo.get("altitud").asLong());
+                                        estacionesDTO.setAltitud(nodo.get("altitud").asShort());
                                         estacionesDTO.setIndicativo(nodo.get("indicativo").asString());
                                         estacionesDTO.setNombre(nodo.get("nombre").asString());
                                         estacionesDTO.setIndsinop(nodo.get("indsinop").asString());
